requirements.txt
django-bootstrap4==1.0.1
django-extensions==2.2.1
django==2.2.4
